Consolidate closure wrangling.
